# feat(perf-detector-threshold-configuration) Changed thresholds of extended detectors.