- The idea is to set the values of the new thresholds, in a way that represents the effect of current thresholds.
- Consecutive HTTP:
- - Min time saved threshold for current detector is roughly 2s. The is achieved by parallelizing 3, 1s spans. 
- N+1 API Calls:
- - Min total duration for current detector is 500ms. Offending list consists of 10, 50ms spans. 
- The loosening of existing defaults will be handled separately.